alpha_validator_mvp/requirements.txt
alphalens-reloaded>=0.4.6
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.0
seaborn>=0.13.2
scipy>=1.11.0
statsmodels>=0.14.1
